mapclient/eth: Format code.

+++ mapclients/eth/contracts/interface/IBLSPoint.sol
@@ -4,13 +4,13 @@ pragma solidity 0.8.7;
 
 interface IBLSPoint {
     struct G1 {
-        uint x;
-        uint y;
+        uint256 x;
+        uint256 y;
     }
     struct G2 {
-        uint xr;
-        uint xi;
-        uint yr;
-        uint yi;
+        uint256 xr;
+        uint256 xi;
+        uint256 yr;
+        uint256 yi;
     }
 }
